move Flow "scripts" and "(un)install" over
