+++ lib/tasks/flowtype-uninstall.js
@@ -5,32 +5,10 @@ const path = require('path');
 
 const pify = require('pify');
 const rimraf = require('rimraf');
-const updateJsonFile = require('update-json-file');
 
-const { removeScript } = require('../pkg.js');
-const { JSON_OPTIONS, PACKAGE_JSON } = require('../values.js');
+const { PACKAGE_JSON } = require('../values.js');
 
-const LABEL = 'unused FlowType uninstalled and configured';
-
-function npmScript(cwd /* : string */) {
-  return updateJsonFile(
-    path.join(cwd, 'package.json'),
-    pkg => {
-      pkg.scripts = pkg.scripts || {};
-
-      for (const script of ['pretest', 'test', 'posttest']) {
-        removeScript(pkg, script, 'npm run flow');
-        removeScript(pkg, script, 'npm run flow_check');
-      }
-
-      delete pkg.scripts.flow;
-      delete pkg.scripts.flow_check;
-
-      return pkg;
-    },
-    JSON_OPTIONS,
-  );
-}
+const LABEL = 'unused FlowType de-configured';
 
 /* :: import type { TaskOptions } from '../types.js' */
 
@@ -45,7 +23,7 @@ function deleteFlowTyped(cwd /* : string */) {
 }
 
 function fn({ cwd } /* : TaskOptions */) {
-  return Promise.all([deleteConfig(cwd), deleteFlowTyped(cwd), npmScript(cwd)]);
+  return Promise.all([deleteConfig(cwd), deleteFlowTyped(cwd)]);
 }
 
 function isNeeded({ isFlowProject } /* : TaskOptions */) {
