+++ lib/tasks/flowtype.js
@@ -3,15 +3,12 @@
 
 const path = require('path');
 
-const updateJsonFile = require('update-json-file');
-
 const { copyFile } = require('../fs.js');
-const { addScript } = require('../pkg.js');
-const { JSON_OPTIONS, PACKAGE_JSON } = require('../values.js');
+const { PACKAGE_JSON } = require('../values.js');
 
 /* :: import type { TaskOptions } from '../types.js' */
 
-const LABEL = 'FlowType installed and configured';
+const LABEL = 'FlowType configured';
 
 const SKELETON_FLOWCONFIG_PATH = path.join(
   __dirname,
@@ -20,26 +17,8 @@ const SKELETON_FLOWCONFIG_PATH = path.join(
   '.flowconfig',
 );
 
-function npmScript(cwd /* : string */) {
-  return updateJsonFile(
-    path.join(cwd, 'package.json'),
-    pkg => {
-      pkg.scripts = pkg.scripts || {};
-
-      pkg.scripts.flow = `flow check`;
-      addScript(pkg, 'test', 'npm run flow');
-
-      return pkg;
-    },
-    JSON_OPTIONS,
-  );
-}
-
-function fn({ cwd } /* : TaskOptions */) {
-  return Promise.all([
-    copyFile(SKELETON_FLOWCONFIG_PATH, path.join(cwd, '.flowconfig')),
-    npmScript(cwd),
-  ]);
+async function fn({ cwd } /* : TaskOptions */) {
+  await copyFile(SKELETON_FLOWCONFIG_PATH, path.join(cwd, '.flowconfig'));
 }
 
 function isNeeded({ isFlowProject } /* : TaskOptions */) {
@@ -51,8 +30,5 @@ module.exports = {
   id: path.basename(__filename),
   isNeeded,
   label: LABEL,
-  lib: {
-    npmScript,
-  },
   requires: [PACKAGE_JSON],
 };
