+++ lib/tasks/npm-scripts.js
@@ -45,6 +45,29 @@ const fmt = () => ({
   'sort-package-json': 'npx -q sort-package-json',
 });
 
+function flow(pkg /* : Object */, isFlowProject /* : boolean */) {
+  if (isFlowProject) {
+    pkg.scripts.flow = `flow check`;
+    addScript(pkg, 'lint', 'npm run flow');
+  } else {
+    pkg.scripts = omit(pkg.scripts, ['flow']);
+    for (const script of ['lint', 'pretest', 'test', 'posttest']) {
+      removeScript(pkg, script, 'npm run flow');
+    }
+  }
+
+  // we call as part of "lint" now, not in "test"
+  removeScript(pkg, 'test', 'npm run flow');
+
+  // drop our old Flow script name
+  pkg.scripts = omit(pkg.scripts, ['flow_check']);
+  for (const script of ['pretest', 'test', 'posttest']) {
+    removeScript(pkg, script, 'npm run flow_check');
+  }
+
+  return pkg;
+}
+
 const lint = () => ({
   eslint: 'eslint --cache --fix .',
   lint: 'npm run eslint',
@@ -85,7 +108,7 @@ function trailingComma(pkg /* : Object */) /* : 'all' | 'es5' */ {
   return 'all';
 }
 
-function fn({ cwd } /* : TaskOptions */) {
+function fn({ cwd, isFlowProject = false } /* : TaskOptions */) {
   return updateJsonFile(
     path.join(cwd, 'package.json'),
     pkg => {
@@ -112,16 +135,14 @@ function fn({ cwd } /* : TaskOptions */) {
         ...lint(),
       };
       addScript(pkg, 'test', 'npm run lint');
+      pkg = flow(
+        pkg,
+        isFlowProject,
+      );
 
       // drop fixpack
       pkg.scripts = omit(pkg.scripts, ['fixpack']);
       removeScript(pkg, 'posttest', 'npm run fixpack');
-
-      // drop our old Flow script name
-      pkg.scripts = omit(pkg.scripts, ['flow_check']);
-      for (const script of ['test', 'posttest']) {
-        removeScript(pkg, script, 'npm run flow_check');
-      }
 
       // stop calling ESLint in the old places / ways
       for (const script of ['test', 'posttest']) {
